Add Documentation To CitiesController

+++ TABP.Application/Handlers/Cities/DeleteCityHandler.cs
@@ -1,10 +1,11 @@
 using MediatR;
 using TABP.Application.Commands.Cities;
+using TABP.Application.Responses;
 using TABP.Domain.Interfaces.Repositories;
 
 namespace TABP.Application.Handlers.Cities;
 
-public class DeleteCityHandler : IRequestHandler<DeleteCityCommand>
+public class DeleteCityHandler : IRequestHandler<DeleteCityCommand, Result>
 {
     private readonly ICityRepository _cityRepository;
 
@@ -13,8 +14,15 @@ public class DeleteCityHandler : IRequestHandler<DeleteCityCommand>
         _cityRepository = cityRepository;
     }
 
-    public async Task Handle(DeleteCityCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteCityCommand request, CancellationToken cancellationToken)
     {
+        var city = await _cityRepository.GetByIdAsync(request.CityId);
+        if (city == null)
+        {
+            return Result.Fail("City not found.");
+        }
         await _cityRepository.DeleteAsync(request.CityId);
+        
+        return Result.Success();
     }
 }
